Restructured root directory to organize files. Some progress on employeeView
