+++ Controllers/error.php
@@ -5,7 +5,7 @@
 <div class="form">
   <h1>Error</h1>
   <?php echo $_SESSION['message']; ?>
-  <a href="index.php"><button class="button button-block"/>Home</button></a>
+  <a href="../index.php"><button class="button button-block"/>Home</button></a>
 </div>
 </body>
 </html>
