+++ registerPage.php
@@ -4,7 +4,7 @@ require 'header.php';
 //If form is post, it'll call itself and run the php code at top w/ method POST.
 if ($_SERVER['REQUEST_METHOD']=='POST') {
     if (isset($_POST['register'])) {
-      require 'registerService.php';
+      require 'Controllers/registerService.php';
     }
 }
 ?>
